refactor: introduce service layer
