+++ Konosuba/AppDelegate.swift
@@ -14,7 +14,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_: UIApplication, didFinishLaunchingWithOptions _: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         let viewController = window?.rootViewController as! ListViewController
-        viewController.reactor = ListViewReactor()
+        let provider = ServiceProvider()
+        viewController.reactor = ListViewReactor(provider: provider)
         return true
     }
