+++ Konosuba/View/ListViewReactor.swift
@@ -27,9 +27,11 @@ final class ListViewReactor: Reactor {
     }
 
     let initialState: State
+    let provider: ServiceProviderType
 
-    init() {
+    init(provider: ServiceProviderType) {
         initialState = State(proposals: [], isLoading: false)
+        self.provider = provider
     }
 
     func mutate(action: Action) -> Observable<Mutation> {
@@ -37,7 +39,7 @@ final class ListViewReactor: Reactor {
         case .initialView, .refresh:
             return Observable.concat([
                 Observable.just(Mutation.setLoading(true)),
-                search().map { Mutation.updateList($0) },
+                provider.proposalService.fetch().map { Mutation.updateList($0) },
                 Observable.just(Mutation.setLoading(false)),
             ])
         }
@@ -52,15 +54,5 @@ final class ListViewReactor: Reactor {
             state.isLoading = loading
         }
         return state
-    }
-
-    func search() -> Observable<[Proposal]> {
-        let url = URL(string: "https://iosdc-cfps.penginmura.tech/api/v1/proposals")!
-        return URLSession.shared.rx.data(request: URLRequest(url: url))
-            .map { data -> [Proposal] in
-                let proposals = try JSONDecoder().decode(ProposalList.self, from: data)
-                return proposals
-            }
-        // TODO: エラー処理
     }
 }
